requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymysql
python-dotenv==1.0.0
cryptography
python-multipart
boto3==1.29.3
requests==2.31.0
scenedetect==0.6.2
opencv-python==4.8.1.78